Add distillery
I abandoned the edeliver way. It doesn't work out of the box and it
seems an unnecesary complex tool.

How does distillery work?

I created a rel/config.exs which sets up all the params needed by
distillery (created by using the mix task that comes with distillery
`mix release.init` )

Once configured you can run:

- `MIX_ENV=prod mix release` which compiles and generates a release on
  _build/prod/
- You can upload the release to the server (there are two options:
  uploading a tar or just rsyncing the files in /release -the latter
  should be faster because it just uploads deltas)
- On the server you run `PORT=8080 .../bin/phoenix_chat start`

I set up distillery to include erts (Erlang Runtime System), so it's
not necessary to have Erlang or mix installed in the final server.

When using this option is not possible to use hot upgrades. If we
want to be able to create .appup upgrades we have to disable this
option.

For the moment, I'll go with the simplest approach: just restart the
server when deploying.

--
I also modified phoenix config/prod.exs:

- commented out cache_static_manifest: "priv/static/manifest.json as I don't care about
assets yet.

- and added server:true which means that the executable should start
the server (listening on the port ready from :system, :port). If you
don't set up the PORT env variable the executable will not start