Distillery - Set up release system
There are two standard options to make releases in Elixir.
The old one is called exrm. The newer one is distillery.

Aside that, there are tools that work on top of a release maker, to
automatize the process.

Our first idea was to use one of the latter. We tried edeliver which
is kind of similar to capistrano in the Ruby world. It works on top
of distillery, it makes releases and then copy the files to the servers
via ssh (the same idea as cap deploy in capistrano). We tried that, but
it didn't work out of the box, we spent some time and decided that it
wasn't worthy. Moreover, I don't like the philosophy of Capistrano, so
we moved to the next option:

· Learn how distillery works

Actually, it's pretty easy. We added distillery as a dependency on our
mix.exs, `mix deps.get` to retrieve the dependencies.
`mix release.init` to create the release config file.

And `MIX_ENV=prod mix release`. Ideally this command should be run in
the build server, with the same OS version of the production server.
This command generates a build (a tar) that can be synchronized to the
production server via rsync or similar.

There are many options in the release config file (rel/config.exs). One
of the important decisions is wheter we want to support hot upgrades. Hot
upgrades allow you to make changes in production without shutting down
servers. Instead of creating a new release, we create an .appup (app
upgrade) which replaces modules on the fly. We found that at this moment
this feature is too complex and we don't need it (at least yet).

One of the advantage of creating a new release each time is that we don't need
to install erlang on the production server. We just pack erts (erlang
run time system) in the release.

And that's it.

When you run `mix release` it says something like:

==> Release successfully built!
    You can run it in one of the following ways:
      Interactive: _build/prod/rel/slackbot/bin/slackbot console
      Foreground: _build/prod/rel/slackbot/bin/slackbot foreground
      Daemon: _build/prod/rel/slackbot/bin/slackbot start

One thing to remember:
The env variables (needed to run, for instance
`Application.get_env(:slackbot, :slack_token)` should be present in the
build server at build time.